# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Docker Image Sync Platform API",
    "description": "Docker镜像同步平台的RESTful API文档",
    "version": "1.0.0",
    "contact": {
      "name": "Docker Image Sync Platform Team",
      "email": "support@example.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
    "/sync/submit": {
      "post": {
        "tags": ["Sync"],
        "summary": "提交单个镜像同步任务",
        "description": "提交一个Docker镜像同步任务，将镜像同步到阿里云ACR",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/SyncRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "同步任务提交成功",
            "schema": { "$ref": "#/definitions/TaskStatusResponse" }
          },
          "400": { "description": "请求参数错误" },
          "500": { "description": "服务器内部错误" }
        }
      }
    },
    "/sync/batch": {
      "post": {
        "tags": ["Sync"],
        "summary": "提交批量镜像同步任务",
        "description": "提交多个Docker镜像的批量同步任务",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/BatchSyncRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "批量同步任务提交成功",
            "schema": { "$ref": "#/definitions/BatchSyncResponse" }
          },
          "400": { "description": "请求参数错误" },
          "500": { "description": "服务器内部错误" }
        }
      }
    },
    "/sync/batch/mock": {
      "post": {
        "tags": ["Sync"],
        "summary": "提交模拟批量同步任务",
        "description": "提交模拟批量同步任务，用于测试和演示，不执行实际镜像同步",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/BatchSyncRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "模拟批量同步任务提交成功",
            "schema": { "$ref": "#/definitions/BatchSyncResponse" }
          }
        }
      }
    },
    "/sync/status/{taskId}": {
      "get": {
        "tags": ["Sync"],
        "summary": "查询单个同步任务状态",
        "description": "根据任务ID查询单个同步任务的执行状态和进度",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "任务状态查询成功",
            "schema": { "$ref": "#/definitions/TaskStatusResponse" }
          },
          "404": { "description": "任务不存在" }
        }
      }
    },
    "/sync/batch/status/{taskId}": {
      "get": {
        "tags": ["Sync"],
        "summary": "查询批量同步任务状态",
        "description": "根据任务ID查询批量同步任务的整体状态和各子任务执行情况",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "批量任务ID"
          }
        ],
        "responses": {
          "200": {
            "description": "批量任务状态查询成功",
            "schema": { "$ref": "#/definitions/BatchStatusResponse" }
          },
          "404": { "description": "任务不存在" }
        }
      }
    },
    "/sync/history": {
      "get": {
        "tags": ["Sync"],
        "summary": "获取同步历史记录",
        "description": "分页查询历史同步任务，支持状态筛选和时间范围查询",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "description": "页码"
          },
          {
            "name": "pageSize",
            "in": "query",
            "type": "integer",
            "default": 10,
            "description": "每页数量"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "enum": ["pending", "running", "success", "failed", "retrying", "skipped"],
            "description": "任务状态筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "同步历史记录获取成功",
            "schema": { "$ref": "#/definitions/SyncHistoryResponse" }
          }
        }
      }
    },
    "/sync/suggest-acr": {
      "get": {
        "tags": ["Sync"],
        "summary": "根据源镜像建议目标 ACR",
        "description": "按仓库名查询历史归属并返回 ACR 建议与配额摘要",
        "parameters": [
          {
            "name": "image",
            "in": "query",
            "type": "string",
            "required": true,
            "description": "源镜像地址，如 nginx:1.21"
          }
        ],
        "responses": {
          "200": { "description": "查询成功" },
          "400": { "description": "请求参数错误" },
          "500": { "description": "服务器内部错误" }
        }
      }
    },
    "/sync/check-acr": {
      "post": {
        "tags": ["Sync"],
        "summary": "批量检查镜像与所选 ACR 的归属冲突",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/CheckAcrRequest" }
          }
        ],
        "responses": {
          "200": { "description": "检查成功" },
          "400": { "description": "请求参数错误" },
          "500": { "description": "服务器内部错误" }
        }
      }
    },
    "/images/list": {
      "get": {
        "tags": ["Images"],
        "summary": "获取镜像列表",
        "description": "获取已同步的镜像列表，支持分页、搜索、状态和架构筛选",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "description": "页码"
          },
          {
            "name": "pageSize",
            "in": "query",
            "type": "integer",
            "default": 10,
            "description": "每页数量"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string",
            "description": "搜索关键词（镜像名称）"
          },
          {
            "name": "architecture",
            "in": "query",
            "type": "string",
            "enum": ["amd64", "arm64"],
            "description": "架构筛选"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "enum": ["pending", "syncing", "success", "failed", "retrying", "skipped"],
            "description": "同步状态筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "镜像列表获取成功",
            "schema": { "$ref": "#/definitions/ImageListResponse" }
          }
        }
      }
    },
...
    "/images/batch-check": {
      "post": {
        "tags": ["Images"],
        "summary": "批量检查镜像是否存在于ACR",
        "description": "批量检查多个镜像是否已同步到阿里云ACR",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/BatchCheckRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "批量检查结果",
            "schema": { "$ref": "#/definitions/BatchCheckResponse" }
          }
        }
      }
    },
    "/images/{id}": {
      "get": {
        "tags": ["Images"],
        "summary": "获取镜像详情",
        "description": "根据ID获取单个镜像的详细同步信息",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "镜像记录ID"
          }
        ],
        "responses": {
          "200": {
            "description": "镜像详情获取成功",
            "schema": { "$ref": "#/definitions/ImageSyncRecord" }
          },
          "404": { "description": "镜像不存在" }
        }
      },
      "delete": {
        "tags": ["Images"],
        "summary": "删除镜像记录",
        "description": "根据ID删除镜像记录（软删除）",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "镜像记录ID"
          }
        ],
        "responses": {
          "200": { "description": "镜像删除成功" },
          "404": { "description": "镜像不存在" }
        }
      }
    },
    "/images/{id}/retry": {
      "post": {
        "tags": ["Images"],
        "summary": "重试镜像同步",
        "description": "对失败的镜像同步任务重新执行",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "镜像记录ID"
          }
        ],
        "responses": {
          "200": { "description": "重试任务提交成功" },
          "404": { "description": "镜像不存在" }
        }
      }
    },
    "/images/{id}/check": {
      "post": {
        "tags": ["Images"],
        "summary": "检查单个镜像是否存在于ACR",
        "description": "检查指定镜像是否已同步到阿里云ACR",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "镜像记录ID"
          }
        ],
        "responses": {
          "200": {
            "description": "检查结果",
            "schema": { "$ref": "#/definitions/ImageCheckResponse" }
          }
        }
      }
    },
    "/github/runs": {
      "get": {
        "tags": ["GitHub"],
        "summary": "获取GitHub Actions工作流运行列表",
        "description": "获取GitHub Actions工作流的运行历史记录",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "default": 1,
            "description": "页码"
          },
          {
            "name": "pageSize",
            "in": "query",
            "type": "integer",
            "default": 10,
            "description": "每页数量"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string",
            "enum": ["queued", "in_progress", "completed"],
            "description": "运行状态筛选"
          }
        ],
        "responses": {
          "200": {
            "description": "工作流运行列表获取成功",
            "schema": { "$ref": "#/definitions/GitHubRunsResponse" }
          }
        }
      }
    },
    "/github/runs/{runId}": {
      "get": {
        "tags": ["GitHub"],
        "summary": "获取GitHub Actions工作流运行详情",
        "description": "根据运行ID获取工作流的详细执行信息和作业日志",
        "parameters": [
          {
            "name": "runId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "工作流运行ID"
          }
        ],
        "responses": {
          "200": {
            "description": "工作流运行详情获取成功",
            "schema": { "$ref": "#/definitions/GitHubRunDetailResponse" }
          }
        }
      }
    },
...
    "/config/git-repository": {
      "get": {
        "tags": ["Config"],
        "summary": "获取Git仓库类型配置",
        "description": "获取当前选择的Git仓库类型（gitee 或 github）",
        "responses": {
          "200": {
            "description": "Git仓库类型获取成功",
            "schema": { "$ref": "#/definitions/GitRepositoryConfigResponse" }
          }
        }
      },
      "put": {
        "tags": ["Config"],
        "summary": "更新Git仓库类型配置",
        "description": "更新Git仓库类型选择（gitee 或 github）",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/GitRepositoryConfigRequest" }
          }
        ],
        "responses": {
          "200": { "description": "Git仓库类型更新成功" },
          "400": { "description": "请求参数错误" }
        }
      }
    },
...
    "/config/git/gitee": {
      "put": {
        "tags": ["Config"],
        "summary": "更新Gitee配置",
        "description": "更新Gitee仓库的配置信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/GiteeConfigRequest" }
          }
        ],
        "responses": {
          "200": { "description": "Gitee配置更新成功" },
          "400": { "description": "请求参数错误" }
        }
      }
    },
    "/config/git/github": {
      "put": {
        "tags": ["Config"],
        "summary": "更新GitHub配置",
        "description": "更新GitHub仓库的配置信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/GitHubConfigRequest" }
          }
        ],
        "responses": {
          "200": { "description": "GitHub配置更新成功" },
          "400": { "description": "请求参数错误" }
        }
      }
    },
    "/config/git/test": {
      "post": {
        "tags": ["Config"],
        "summary": "测试Git连接",
        "description": "测试Git仓库的连接是否正常，验证认证信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/GitTestRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "Git连接测试结果",
            "schema": { "$ref": "#/definitions/GitTestResponse" }
          }
        }
      }
    },
    "/config/git-test-operations": {
      "post": {
        "tags": ["Config"],
        "summary": "测试Git代码拉取和提交操作",
        "description": "完整测试GitHub代码拉取、提交和推送三步操作，用于验证配置正确性",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/GitTestOperationsRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "测试操作结果",
            "schema": { "$ref": "#/definitions/GitTestOperationsResponse" }
          }
        }
      }
    },
    "/config/git-optimization": {
      "get": {
        "tags": ["Config"],
        "summary": "获取Git优化配置",
        "description": "获取Git操作的优化参数配置（超时、缓存、稀疏检出等）",
        "responses": {
          "200": {
            "description": "Git优化配置获取成功",
            "schema": { "$ref": "#/definitions/GitOptimizationConfigResponse" }
          }
        }
      },
      "put": {
        "tags": ["Config"],
        "summary": "更新Git优化配置",
        "description": "更新Git操作的优化参数配置",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/GitOptimizationConfigRequest" }
          }
        ],
        "responses": {
          "200": { "description": "Git优化配置更新成功" }
        }
      }
    },
...
    "/config/aliyun-db": {
      "get": {
        "tags": ["Config"],
        "summary": "获取阿里云ACR配置",
        "description": "获取阿里云容器镜像服务的配置信息（数据库版本）",
        "responses": {
          "200": {
            "description": "阿里云配置获取成功",
            "schema": { "$ref": "#/definitions/AliyunConfigResponse" }
          }
        }
      },
      "put": {
        "tags": ["Config"],
        "summary": "更新阿里云ACR配置",
        "description": "更新阿里云容器镜像服务的配置信息",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/AliyunConfigRequest" }
          }
        ],
        "responses": {
          "200": { "description": "阿里云配置更新成功" },
          "400": { "description": "请求参数错误" }
        }
      }
    },
    "/config/aliyun/test": {
      "post": {
        "tags": ["Config"],
        "summary": "测试阿里云ACR连接",
        "description": "测试阿里云容器镜像服务的连接是否正常",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": { "$ref": "#/definitions/AliyunTestRequest" }
          }
        ],
        "responses": {
          "200": {
            "description": "阿里云连接测试结果",
            "schema": { "$ref": "#/definitions/AliyunTestResponse" }
          }
        }
      }
    },
...
    "/config/debug/{key}": {
      "get": {
        "tags": ["Config"],
        "summary": "调试获取指定配置项",
        "description": "根据配置键名获取单个配置值（调试用途）",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "配置键名"
          }
        ],
        "responses": {
          "200": { "description": "配置项获取成功" },
          "404": { "description": "配置项不存在" }
        }
      }
    },
...
    "/auth/roles": {
      "get": {
        "tags": ["Auth"],
        "summary": "角色列表",
        "description": "获取所有角色及其权限（需 roles 权限）",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "成功" },
          "403": { "description": "无权限" }
        }
      },
      "post": {
        "tags": ["Auth"],
        "summary": "创建角色",
        "description": "创建自定义角色（需 roles 权限）",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{
          "in": "body",
          "name": "body",
          "required": true,
          "schema": { "$ref": "#/definitions/CreateRoleRequest" }
        }],
        "responses": {
          "200": { "description": "创建成功" },
          "400": { "description": "参数错误" },
          "403": { "description": "无权限" }
        }
      }
    },
    "/auth/roles/{id}": {
      "get": {
        "tags": ["Auth"],
        "summary": "角色详情",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "成功" }, "404": { "description": "不存在" } }
      },
      "put": {
        "tags": ["Auth"],
        "summary": "更新角色",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "integer" },
          { "in": "body", "name": "body", "required": true, "schema": { "$ref": "#/definitions/UpdateRoleRequest" } }
        ],
        "responses": { "200": { "description": "更新成功" }, "400": { "description": "参数错误" } }
      },
      "delete": {
        "tags": ["Auth"],
        "summary": "删除角色",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "type": "integer" }],
        "responses": { "200": { "description": "删除成功" }, "400": { "description": "不可删除" } }
      }
    },
    "/auth/permissions": {
      "get": {
        "tags": ["Auth"],
        "summary": "权限列表",
        "description": "获取系统注册的全部权限（需 roles 权限）",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "成功" } }
      }
    },
    "/auth/roles/options": {
      "get": {
        "tags": ["Auth"],
        "summary": "角色下拉选项",
        "description": "供用户管理表单使用（仅需登录）",
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "成功" } }
      }
    }
  },
  "definitions": {
    "SyncRequest": {
      "type": "object",
      "required": ["images"],
      "properties": {
        "images": {
          "type": "array",
          "items": { "type": "string" },
          "description": "要同步的Docker镜像列表，例如 nginx:latest"
        },
        "architecture": {
          "type": "string",
          "enum": ["amd64", "arm64"],
          "default": "amd64",
          "description": "目标CPU架构"
        },
        "description": {
          "type": "string",
          "description": "同步任务描述"
        }
      }
    },
...
    "TaskStatusResponse": {
      "type": "object",
      "properties": {
        "task_id": { "type": "string", "description": "任务ID" },
        "status": {
          "type": "string",
          "enum": ["pending", "running", "success", "failed", "retrying", "skipped"],
          "description": "任务状态"
        },
        "source_image": { "type": "string", "description": "源镜像名称" },
        "target_image": { "type": "string", "description": "目标镜像名称（ACR地址）" },
        "architecture": { "type": "string", "description": "目标CPU架构" },
        "github_action_url": { "type": "string", "description": "GitHub Actions工作流URL" },
        "started_at": { "type": "string", "format": "date-time", "description": "任务开始时间" },
        "completed_at": { "type": "string", "format": "date-time", "description": "任务完成时间" },
        "error_message": { "type": "string", "description": "错误信息（如有）" },
        "images": {
          "type": "array",
          "items": { "type": "string" },
          "description": "涉及的镜像列表"
        }
      }
    },
...
  }
}